output/requirements.txt
google-genai>=1.0.0
pydantic>=2.0
pydantic-settings>=2.0
aiohttp>=3.9
pyyaml>=6.0
rich>=13.0
